+++ et/audio/modules/op.cc
@@ -3,7 +3,7 @@
 namespace Et {
 namespace Audio {
 
-Op::Op(unsigned int sampleRate, unsigned int bufferSize)
+Op::Op(unsigned int sampleRate, size_t bufferSize)
     : Module(sampleRate, bufferSize, inputCount, parameterCount)
     , osc(sampleRate, bufferSize)
     , adsr(sampleRate, bufferSize)
@@ -15,7 +15,7 @@ void Op::process()
 {
     adsr.tick(lastSampleId_);
     
-    for(unsigned int i=0; i<bufferSize_; ++i) {
+    for(size_t i=0; i<bufferSize_; ++i) {
         output_.setSample(Buffer::Channel::Left, i, 
             adsr.getOutput().getSample(Buffer::Channel::Left, i));
         output_.setSample(Buffer::Channel::Right, i,
